+++ client/projects/cms/src/app/modules/dashboard/modules/module-instance/pages/instance-single/instance-single.component.ts
@@ -1,6 +1,5 @@
 import {ChangeDetectionStrategy, Component, Injector, OnInit, ViewChild} from '@angular/core';
 import {ActivatedRoute, Router} from '@angular/router';
-import {ModuleAuthorization} from 'definitions';
 import {FormBuilderData, FormBuilderComponent, State, FormBuilderContextService} from '@jaspero/form-builder';
 import {parseTemplate, random, safeEval} from '@jaspero/utils';
 import {UntilDestroy, untilDestroyed} from '@ngneat/until-destroy';
@@ -33,7 +32,6 @@ interface Instance {
   formatOnSave: (data: any) => any;
   formatOnEdit: (data: any) => any;
   formatOnCreate: (data: any) => any;
-  authorization?: ModuleAuthorization;
   formBuilder: FormBuilderData;
 }
 
@@ -129,8 +127,7 @@ export class InstanceSingleComponent implements OnInit {
             this.confirmExitOnTouched = module.metadata?.hasOwnProperty('confirmExitOnTouched')
               && module.metadata?.confirmExitOnTouched && this.currentState === ViewState.Edit;
 
-            if (module.layout) {
-              if (module.layout.editTitleKey) {
+              if (module.layout?.editTitleKey) {
                 const evaluated = safeEval(module.layout.editTitleKey);
                 const eValue = (
                   typeof evaluated === 'function'
@@ -140,8 +137,11 @@ export class InstanceSingleComponent implements OnInit {
                 editTitleKey = eValue !== undefined && eValue !== 'undefined' ?
                   eValue :
                   (module.layout.editTitleKeyFallback || '-');
+              } else {
+                editTitleKey = value[editTitleKey];
               }
 
+            if (module.layout) {
               if (module.layout.instance) {
 
                 if (module.layout.instance.actions) {
@@ -195,7 +195,6 @@ export class InstanceSingleComponent implements OnInit {
                 editTitleKey
               },
               directLink: !!(module.layout && module.layout.directLink),
-              authorization: module.authorization,
               ...actions && {actions},
               ...formatOn,
               formBuilder: {