+++ client/projects/cms/src/app/shared/modules/fb/form-builder-shared.module.ts
@@ -1,6 +1,8 @@
 import {CommonModule} from '@angular/common';
 import {NgModule} from '@angular/core';
+import {ReactiveFormsModule} from '@angular/forms';
 import {MatButtonModule} from '@angular/material/button';
+import {MatCheckboxModule} from '@angular/material/checkbox';
 import {MatIconModule} from '@angular/material/icon';
 import {FbFieldsMatModule} from '@jaspero/fb-fields-mat';
 import {FB_PAGE_BUILDER_OPTIONS, PageBuilderModule} from '@jaspero/fb-page-builder';
@@ -13,13 +15,14 @@ import {
 } from '@jaspero/form-builder';
 import {TranslocoModule} from '@ngneat/transloco';
 import {TinymceModule} from '@jaspero/fb-tinymce';
-import {BlocksModule} from '@shared/blocks/blocks.module';
 import {FbStorageService} from '../../../../../integrations/firebase/fb-storage.service';
 import {environment} from '../../../../environments/environment';
 import {DbService} from '../../services/db/db.service';
 import {StateService} from '../../services/state/state.service';
 import {DuplicateComponent} from './custom-components/duplicate/duplicate.component';
+import {BlocksModule} from '@shared/blocks/blocks.module';
 import {EmailTemplateDescriptionComponent} from './custom-components/email-template-description/email-template-description.component';
+import {PermissionsComponent} from './custom-fields/permissions/permissions.component';
 
 export function roleFactory(state: StateService) {
   return state.role;
@@ -28,17 +31,21 @@ export function roleFactory(state: StateService) {
 @NgModule({
   imports: [
     CommonModule,
+    ReactiveFormsModule,
+
     FormBuilderModule.forRoot(),
     FbFieldsMatModule.forRoot({prefix: ''}),
     FbSegmentsMatModule.forRoot({prefix: ''}),
-    PageBuilderModule,
     TinymceModule,
+    PageBuilderModule,
 
     /**
      * Custom fields and component dependencies
      */
     MatButtonModule,
     MatIconModule,
+    MatCheckboxModule,
+
     TranslocoModule
   ],
   exports: [
@@ -71,7 +78,9 @@ export function roleFactory(state: StateService) {
     },
     {
       provide: CUSTOM_FIELDS,
-      useValue: {}
+      useValue: {
+        permissions: PermissionsComponent
+      }
     },
     {
       provide: FB_PAGE_BUILDER_OPTIONS,
@@ -81,8 +90,17 @@ export function roleFactory(state: StateService) {
     },
   ],
   declarations: [
+
+    /**
+     * Custom Components
+     */
     DuplicateComponent,
-    EmailTemplateDescriptionComponent
+    EmailTemplateDescriptionComponent,
+
+    /**
+     * Custom Fields
+     */
+    PermissionsComponent
   ]
 })
-export class FormBuilderSharedModule {}
+export class FormBuilderSharedModule { }