+++ functions/src/callable/update-email.ts
@@ -1,15 +1,29 @@
+import {Collections, SHARED_CONFIG} from 'definitions';
 import {auth, firestore} from 'firebase-admin';
 import * as functions from 'firebase-functions';
-import {SHARED_CONFIG, Collections} from 'definitions';
-import {hasRole} from '../utils/auth';
+import {hasPermission} from '../utils/auth';
+import {schemaValidation} from '../utils/schema-validation';
 
 export const updateEmail = functions
   .region(SHARED_CONFIG.cloudRegion)
   .https.onCall(async (data, context) => {
-    hasRole(context, 'admin');
+    hasPermission(context, Collections.Users, 'update');
+    schemaValidation(
+      data,
+      {
+        additionalProperties: true,
+        properties: {
+          id: {type: 'string'},
+          email: {type: 'string'}
+        },
+        required: [
+          'id',
+          'email'
+        ]
+      }
+    );
 
-    const id = data.id;
-    const email = data.email;
+    const {id, email} = data;
 
     try {
       await auth().updateUser(id, {email});
@@ -17,7 +31,7 @@ export const updateEmail = functions
         email
       });
     } catch (e) {
-      console.error(e);
-      throw new functions.https.HttpsError('internal', e.message);
+      functions.logger.error(e);
+      throw new functions.https.HttpsError('internal', e.toString());
     }
   });