+++ functions/src/callable/update-user.ts
@@ -1,13 +1,13 @@
+import {Collections, SHARED_CONFIG} from 'definitions';
 import {auth} from 'firebase-admin';
 import * as functions from 'firebase-functions';
-import {SHARED_CONFIG} from 'definitions';
-import {hasRole} from '../utils/auth';
+import {hasPermission} from '../utils/auth';
 
 export const updateUser = functions
   .region(SHARED_CONFIG.cloudRegion)
   .https
   .onCall(async (data, context) => {
-    hasRole(context, 'admin');
+    hasPermission(context, Collections.Users, 'update');
 
     const ah = auth();
     const {id, ...update} = data;
@@ -38,7 +38,7 @@ export const updateUser = functions
     try {
       await ah.updateUser(id, update);
     } catch (e) {
-      console.error(e);
-      throw new functions.https.HttpsError('internal', e.message);
+      functions.logger.error(e);
+      throw new functions.https.HttpsError('internal', e.toString());
     }
   });